+++ src/lib/skk/editor/IEditor.ts
