+++ src/lib/skk/input-mode/AbstractInputMode.ts
