+++ src/lib/skk/input-mode/IInputMode.ts
