+++ src/lib/skk/input-mode/henkan/AbstractHenkanMode.ts
