+++ src/lib/skk/input-mode/henkan/AbstractMidashigoMode.ts
