+++ src/lib/skk/input-mode/henkan/CandidateDeletionMode.ts
