+++ src/lib/skk/input-mode/henkan/InlineHenkanMode.ts
