+++ src/lib/skk/input-mode/henkan/MenuHenkanMode.ts
