+++ src/lib/skk/jisyo/IJisyoProvider.ts
