+++ src/lib/skk/jisyo/candidate.ts
