+++ src/lib/skk/jisyo/entry.ts
