+++ src/lib/skk/jisyo/jisyo.ts
