+++ src/lib/skk/jisyo/okuri.ts
