+++ src/lib/skk/jisyo/types.ts
